Update dependencies for modern versions (#3)
* Remove dependency for jsonwebtoken

Function is no longer used for this OAuth strategy.

* Remove unused module import

Module is no longer used, so should be removed.

* Update dependencies to more modern versions

Some of these dependencies are running on versions well out of support.

* Increment versions

NodeJS oldest supported version is 14.x, so should be updated to reflect that. Version of package.json should be incremented for these changes.

* Additional dependency update

Dev dependencies also out of date and should be updated to reflect more modern versions.